requirements.txt
av==16.0.1
cffi==2.0.0
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
h5py==3.15.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
kiwisolver==1.4.9
llvmlite==0.45.1
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numba==0.62.1
numpy==2.3.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
pillow==12.0.0
psutil==7.1.3
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
resampy==0.4.3
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
soundfile==0.13.1
sympy==1.14.0
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
triton==3.5.0
typing_extensions==4.15.0
tzdata==2025.2
